fix(bbdo): replace index_id with uint64 type
* fix(conflict_manager_storage): replace index_data in uint64
* enh(status.hh): change type for index_data
* enh(storage:status): index_is is uint64 now.
* enh(bbdo/stream): add ulong in serialize and unserialize + unit test is ok
* enh(conflict_manager): replace index_id in uint64
* enh(conflict_manager): fix troncated index_id in metrics table
* enh(storage): fix syntax

REFS: MON-6801

+++ storage/inc/com/centreon/broker/storage/conflict_manager.hh
@@ -107,7 +107,7 @@ class conflict_manager {
 
   struct index_info {
     std::string host_name;
-    uint32_t index_id;
+    uint64_t index_id;
     bool locked;
     uint32_t rrd_retention;
     std::string service_description;
@@ -181,7 +181,7 @@ class conflict_manager {
   std::unordered_map<uint64_t, size_t> _cache_hst_cmd;
   std::unordered_map<std::pair<uint64_t, uint64_t>, size_t> _cache_svc_cmd;
   std::unordered_map<std::pair<uint64_t, uint64_t>, index_info> _index_cache;
-  std::unordered_map<std::pair<uint32_t, std::string>, metric_info>
+  std::unordered_map<std::pair<uint64_t, std::string>, metric_info>
       _metric_cache;
   std::mutex _metric_cache_m;
